Move StopWordsCriteria to hf_utils.py
